Renaming 'Token' to 'Element'
The term "token" is not always appropriate. The RFC uses the much better term "element". A token is an element that has meaning on its own. An element does not have to be meaningful, but can participate in a sequence of elements that together form a  meaningful token element.

Example:
INT   = 1*DIGIT
DIGIT = %x30-39

In this example, an INT has meaning, so it is a token element. By contrast, each occurrence of DIGIT does not mean anything by itself. We can't derive the completed number, or do any meaningful calculations with a single DIGIT. So any instance of DIGIT is also an element, but not a token.