Change ILexer
Lexers now support ambiguous grammars by branching off at decision points.
Instead of a single best-effort result, all candidate matches are
enumerated.

A candidate match is any string that can be parsed according to the
grammar of the language.

Candidates results are enumerated in a lazy fashion. This usually means that the
first candidate is the shortest and the last candidate is the longest, although I don't know how to prove that.

I have provided two overloads:

 - Read(ITextScanner, ITextContext) : IEnumerable<TElement>

     Enumerates all candidate matches that begin at the character position
     described by the specified ITextContext.

 - Read(ITextScanner) : TElement

     Enumerates all candidate matches that begin at the scanner's current
     position and returns the longest match.